feat(cantidad): el campo "Cant." admite decimales
El campo era digitsOnly: no dejaba ni teclear la coma. Al escribir "2.5"
se comia el punto y guardaba 25 — visto probando en el movil. Y aunque se
hubiera podido teclear, los cuatro sitios que montan el payload enviaban
l.quantity.toInt(), que habria truncado igual.

La columna es numeric(14,3) en toda la base y el escritorio factura
decimales desde siempre, asi que "5,5 horas" o "12,75 metros" eran
posibles alli e imposibles aqui.

- number_parsing.dart: kMaxQuantityDecimals y hasValidQuantityPrecision,
  port del criterio del escritorio (comparacion con tolerancia, porque
  1.1 * 1000 da 1100.0000000000002). Y formatQuantity, que pinta "2" para
  una cantidad entera y "2,5" cuando no lo es.
- El campo acepta coma y punto: el teclado ofrece una u otro segun la
  configuracion del telefono, y parseLocalizedDecimal entiende los dos.
- Validador propio: minimo 0,001 y tope de tres decimales, con el mismo
  mensaje que devuelve la API si se cuela algo.
- Fuera los toInt() de los payloads de factura, presupuesto y venta, y de
  los detalles que pintaban la cantidad.

Afecta a facturas, presupuestos y ventas: comparten el editor de lineas.

Tests: flutter test 88/0 (9 casos nuevos en quantity_test.dart).

+++ lib/features/budgets/create_budget_screen.dart
@@ -93,7 +93,7 @@ class _CreateBudgetScreenState extends ConsumerState<CreateBudgetScreen> {
       'lines': _lines
           .map((l) => {
                 'description': l.description.trim(),
-                'quantity': l.quantity.toInt(),
+                'quantity': l.quantity,
                 'unitPrice': l.unitPrice,
                 'taxRate': l.taxRate,
                 'discountRate': l.discountRate,
